requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==1.4.50
python-dotenv==1.0.0
google-generativeai==0.3.1
openai==1.3.5
requests==2.31.0
python-multipart==0.0.6
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
alembic==1.8.1
# Remove psycopg2-binary since we'll use SQLite
redis==5.0.1
beautifulsoup4==4.12.2
pydantic-settings==2.1.0
jinja2==3.1.2
